+++ src/Service/Yaml/Linter/FileTransformedLinter.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Aeliot\Bundle\TransMaintain\Service\Yaml\Linter;
 
 use Aeliot\Bundle\TransMaintain\Dto\LintYamlFilterDto;
-use Aeliot\Bundle\TransMaintain\Model\FilesTransformedLine;
 use Aeliot\Bundle\TransMaintain\Model\ReportBag;
 use Aeliot\Bundle\TransMaintain\Service\Yaml\FileMapFilter;
 use Aeliot\Bundle\TransMaintain\Service\Yaml\FileTransformedStateDetector;
@@ -38,19 +37,32 @@ final class FileTransformedLinter implements LinterInterface
 
     public function lint(LintYamlFilterDto $filterDto): ReportBag
     {
-        $bag = new ReportBag(FilesTransformedLine::class);
+        $bag = $this->createReportBag();
         $domainsFiles = $this->fileMapFilter->getFilesMap($filterDto);
 
         foreach ($domainsFiles as $domain => $localesFiles) {
             foreach ($localesFiles as $locale => $files) {
                 foreach ($files as $file) {
                     if (!$this->fileTransformedStateDetector->isTransformed($file)) {
-                        $bag->addLine(new FilesTransformedLine($domain, $locale, $file));
+                        $bag->addLine($domain, $locale, $file);
                     }
                 }
             }
         }
 
         return $bag;
+    }
+
+    private function createReportBag(): ReportBag
+    {
+        return new ReportBag(
+            [
+                'domain' => ['string'],
+                'locale' => ['string'],
+                'file' => ['string'],
+            ],
+            'All files have normalised state (they are transformed)',
+            'Files which have abnormal state (they are NOT transformed)'
+        );
     }
 }