+++ src/Service/Yaml/Linter/KeyPatternLinter.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Aeliot\Bundle\TransMaintain\Service\Yaml\Linter;
 
 use Aeliot\Bundle\TransMaintain\Dto\LintYamlFilterDto;
-use Aeliot\Bundle\TransMaintain\Model\KeysPatternLine;
 use Aeliot\Bundle\TransMaintain\Model\ReportBag;
 use Aeliot\Bundle\TransMaintain\Service\Yaml\FileMapFilter;
 use Aeliot\Bundle\TransMaintain\Service\Yaml\KeysParser;
@@ -39,18 +38,31 @@ final class KeyPatternLinter implements LinterInterface
             throw new \LogicException('YAML Key Pattern Linter is not configured yet');
         }
 
-        $bag = new ReportBag(KeysPatternLine::class);
+        $bag = $this->createReportBag();
         $domainsFiles = $this->fileMapFilter->getFilesMap($filterDto);
         foreach ($domainsFiles as $domain => $localesFiles) {
             $keys = $this->getKeysSummary($this->keysParser->getParsedKeys($localesFiles));
             foreach ($keys as $translationId => $locales) {
                 if (!preg_match($this->keyPattern, $translationId)) {
-                    $bag->addLine(new KeysPatternLine($domain, $translationId, $locales));
+                    $bag->addLine($domain, $translationId, $locales);
                 }
             }
         }
 
         return $bag;
+    }
+
+    private function createReportBag(): ReportBag
+    {
+        return new ReportBag(
+            [
+                'domain' => ['string'],
+                'invalid_translation_id' => ['string'],
+                'locales' => ['array'],
+            ],
+            'All translation keys match configured pattern',
+            'Translation keys that are not match configured pattern'
+        );
     }
 
     private function getKeysSummary(array $parsedKeys): array